+++ src/app/profile/actions.ts
@@ -12,6 +12,14 @@ import {
 } from '@/features/profile/constants'
 import { bigIntFix } from '@/lib/bigIntFix'
 import { ErrorResponse } from '@/features/error-handling/types'
+import {
+  Build,
+  BuildItems,
+  BuildReports,
+  BuildVoteCounts,
+  PaidUsers,
+  User,
+} from '@prisma/client'
 
 export type CreatedBuildsFilter = 'date created' | 'upvotes'
 
@@ -116,119 +124,7 @@ export async function getCreatedBuilds({
   })
 }
 
-export type FavoritedBuildsFilter = 'date favorited' | 'upvotes'
 
-export async function getFavoritedBuilds({
-  itemsPerPage,
-  pageNumber,
-  filter,
-}: {
-  itemsPerPage: number
-  pageNumber: number
-  filter: FavoritedBuildsFilter
-}): Promise<PaginationResponse<DBBuild>> {
-  const session = await getServerSession()
-
-  const userId = session?.user?.id
-
-  // find all builds that the user has favorited but are not created
-  // by the user
-  const builds =
-    filter === 'date favorited'
-      ? await prisma.build.findMany({
-          where: {
-            BuildVotes: {
-              some: {
-                userId,
-              },
-            },
-            createdById: {
-              not: userId,
-            },
-          },
-          include: {
-            createdBy: {
-              include: {
-                PaidUsers: true, // Include the related PaidUsers record
-              },
-            },
-            BuildVotes: true,
-            BuildReports: true,
-            BuildItems: true,
-          },
-          skip: (pageNumber - 1) * itemsPerPage,
-          take: itemsPerPage,
-        })
-      : await prisma.build.findMany({
-          where: {
-            BuildVotes: {
-              some: {
-                userId,
-              },
-            },
-            createdById: {
-              not: userId,
-            },
-          },
-          orderBy: {
-            BuildVotes: {
-              _count: 'desc',
-            },
-          },
-          include: {
-            createdBy: {
-              include: {
-                PaidUsers: true, // Include the related PaidUsers record
-              },
-            },
-            BuildVotes: true,
-            BuildReports: true,
-            BuildItems: true,
-          },
-          skip: (pageNumber - 1) * itemsPerPage,
-          take: itemsPerPage,
-        })
-
-  // get the total number of builds that match the conditions
-  const totalBuildCount = await prisma.build.count({
-    where: {
-      BuildVotes: {
-        some: {
-          userId,
-        },
-      },
-      createdById: {
-        not: userId,
-      },
-    },
-  })
-
-  if (!builds) return { items: [], totalItemCount: 0 }
-
-  const returnedBuilds: DBBuild[] = builds.map((build) => ({
-    id: build.id,
-    name: build.name,
-    description: build.description,
-    isPublic: build.isPublic,
-    isFeaturedBuild: build.isFeaturedBuild,
-    thumbnailUrl: build.thumbnailUrl,
-    videoUrl: build.videoUrl,
-    createdById: build.createdById,
-    createdAt: build.createdAt,
-    updatedAt: build.updatedAt,
-    createdByDisplayName:
-      build.createdBy?.displayName ||
-      build.createdBy?.name ||
-      DEFAULT_DISPLAY_NAME,
-    totalUpvotes: build.BuildVotes.length, // Count the votes
-    upvoted: build.BuildVotes.some((vote) => vote.userId === userId), // Check if the user upvoted the build
-    reported: build.BuildReports.some((report) => report.userId === userId), // Check if the user reported the build
-    isMember: build.createdBy.PaidUsers.length > 0, // Check if the user is a member
-    buildItems: build.BuildItems,
-  }))
-
-  return bigIntFix({ items: returnedBuilds, totalItemCount: totalBuildCount })
-}
 
 export async function updateUserDisplayName(
   data: string,